# Measure UDP PPS in performance metrics