Pump as many packets as possible through the link. Account for only the packets that are not lost.

Output looks like this:

```
Test 'virtio_net_throughput_single_queue_tx_pps' running .. (control: test_timeout = 10s, test_iterations = 5, num_queues = 2, queue_size = 256, rx = false, bandwidth = false, overrides: )
Test 'virtio_net_throughput_single_queue_tx_pps' .. ok: mean = 495612.1093586601, std_dev = 34035.03688487209
Test 'virtio_net_throughput_single_queue_rx_pps' running .. (control: test_timeout = 10s, test_iterations = 5, num_queues = 2, queue_size = 256, rx = true, bandwidth = false, overrides: )
Test 'virtio_net_throughput_single_queue_rx_pps' .. ok: mean = 237217.9915156765, std_dev = 2578.1319394996667
Test 'virtio_net_throughput_multi_queue_tx_pps' running .. (control: test_timeout = 10s, test_iterations = 5, num_queues = 4, queue_size = 256, rx = false, bandwidth = false, overrides: )
Test 'virtio_net_throughput_multi_queue_tx_pps' .. ok: mean = 530321.4940044105, std_dev = 33546.872191479095
Test 'virtio_net_throughput_multi_queue_rx_pps' running .. (control: test_timeout = 10s, test_iterations = 5, num_queues = 4, queue_size = 256, rx = true, bandwidth = false, overrides: )
Test 'virtio_net_throughput_multi_queue_rx_pps' .. ok: mean = 240654.67634990936, std_dev = 51097.366007567005
```